refactor: unify duplicated withApiHandler and response helpers
